DeepSeekSummarizer/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
